fix(codemods): update icons-react-size-prop to work with inline props
